gatling: remove uint32_t/int32_t typedefs in shaders.
The idea is sound: if extensions provide types like uint8_t, why not
conform to the naming scheme and add equivalents for other types?
There are multiple problems. First, it's very easy to forget to use
the new name, which leads to inconsistencies. Secondly, no one else
does it. Third, the new type name takes longer to type, adds visual
noise and is not supported by common GLSL syntax highlighting. That
is why I'm removing this feature.