gatling: use relaxed atomic semantics instead of acquire for float adds.
Relaxed/acquire/release atomic semantics define how the compiler
can reorder loads and stores related to atomic instructions. In
this case, most variables are thread-local. Furthermore, both
atomic operations atomicLoad and atomicCompSwap operate on the
same memory location, introducing a data dependency. Since there
are no operations on shared/global memory and the atomic
operations cannot be reordered, too, I assume that we can use
the relaxed semantics.
I've disabled ray jittering and rendered the image using 16 spp.
The md5 hash seems to be the same (this could still be dependent
on the hardware, but the logic seems sound).