gp: implement spatial splitting.
Finally, this feature is done! Avoiding floating-point precision
issues without explicit handling or hardcoded biases was quite
painful.
Aside from the actual spatial splitting, I've also implemented
usage of a reverse-buffer for ranges which get too small. Splitting
is performed only when a certain condition is met, which can be
parameterized using an 'alpha'-value as described in the paper.