cmake: add dependecy solution using CMake FetchContent.
CMake FetchContent is a better solution than git submodules because of
multiple reasons: first, a hierarchy exists and the subproject is not
able to pollute the global namespace or change important settings.
Secondly, the user does not have to do a recursive clone.
It differs from ExternalContent in that we don't have two separate
configure phases - we can use the subproject targets like we can do with
git submodules. However, we're not only limited to submodules!