cmake/gi: add option to use DXC instead of glslang.
When glslang compiles HLSL to SPIR-V, it performs no validation. Hence,
the resulting bytecode might not conform to the spec.

DXC can be used by the developer as a reference compiler/validator.

Since it's based on LLVM and the build process for non-Windows platforms
seems a bit wonky, it's not activated by default or integrated into the
project. It must be installed manually and activated using the GATLING_USE_DXC
CMake option.