shadergen: make sure MaterialX to MDL code generation is thread-safe.
Since we translate MaterialX documents to MDL in the Hydra sync stage
to benefit from multi-threading, we have to make sure that there is no
unsynchronized manipulation of shared state.

According to the header file, the generation context (GenContext) is
supposed to be thread-local. We achieve this by reconstructing it every
time translate() is called.