cmake/hdGatling: fix plugInfo.json installation
One of the new dependencies broke the current approach. As a
(temporary?) alternative, we now install the plugInfo.json file like most
render delegates do. The problem with that is, however, that we can't
guarantee that the configured library name is correct. The only way to get
the actual name is with a generator expression, which can't be used in
the configure_file call at build time.

It might be possible to use a custom command for this purpose.