meta: remove editorconfig
The only IDE/editor, that I care about, where it works out-of-the-box,
is Visual Studio. For other applications, plugins need to be installed:
CLion, Sublime, VSCode, Vim, Notepad++. Most people won't have the
plugin installed, and I rather want to see the root dir less polluted
with files. Whitespace can be commented on in PRs or changed after
merges, and if scale becomes a problem, this file can still be
reintroduced.