GA/docs: update to USD v24.03
Gladly, the same version of Python is used across all platforms.
However, since they differ from the Python versions of the previous USD
builds, we will only build and test against v24.03 for a while (and not
older versions). v24.03 is a good baseline version, because Hydra
colorspace transmission and the Storm normalmapping fix is included.