gi: respect double-sidedness for shading
Not sure if correct yet, but we can further refine the logic with more
test scenes.

Based on my understanding of the MaterialX spec

> Note that in order to have surface shaders set for both sides the
> geometry has to be set as double-sided. Geometry sidedness is a property
> not handled by MaterialX and needs to be set elsewhere.

non double-sided geometry should have the same surface shader on both
sides. Implementing this behaviour also results in both the coffeemaker
and 2DGS scene rendering correctly.